ukl: Add basic UKL configuration & musl mock to the kernel
This patches introduces numerous things. First of all, I added the ukl/
subdirectory in the tree, which contains a musl (libc) mock for future
UKL applications. For the moment, I added 4 syscalls, open / read /
write and close. Apart from the O_CLOEXEC option on an open, they should
behave normally.
I also added a new boot option, ukl. When this option is present, and
the kernel is compiled in UKL mode, no userspace init is launched, and
the main of the UKL program is runned.

One should be able to run a little test program with this patch:

    #include <unistd.h>
    #include <fcntl.h>

    #ifndef USERSPACE
    # define main ukl_main
    #endif

    int main(void) {
        char        buf[256] = { 0 };
        int         fd = open("/etc/passwd", O_RDONLY);
        ssize_t     ret;

        if (fd == -1)
            return 1;

        while ((ret = read(fd, buf, sizeof(buf))))
        {
            buf[ret] = 0;
            write(1, buf, ret);
        }

        close(fd);

        return 0;
    }

The compilation is handled the same way kernel modules does:

    export ukl-obj-m := main.o

    all:
    	make -C ../whiterose/ UKL=$(PWD) bzImage

Where 'whiterose' is the kernel tree.
One can then run this kernel under qemu with:

    qemu-system-x86_64 -enable-kvm -m 1G -s -kernel
    ../whiterose/arch/x86/boot/bzImage -append
    "console=ttyS0 root=/dev/sda ukl quiet" -nographic  -hda qemu-image.img

If the disk image do have an /etc/passwd file, it should be printed on
the console.

Signed-off-by: Louis Solofrizzo <louis@ne02ptzero.me>