ukl: Add fcntl function in the library
This patches adds the supports for fcntl(2) in the ukl library. Should
be working the same as the musl. I've also modified open for a corner
case with FD_CLOEXEC.

There was also a bug in close(2), which have been fixed. I've modified
some error messages for stdin, stdout and stderr mocking.

Signed-off-by: Louis Solofrizzo <louis@ne02ptzero.me>