ukl: basic stdio library and lseek, readv and writev
First of all, this patches introduces three new syscalls in UKL: lseek,
readv and writev. All three are needed to implement the following stdio
functions:
- fopen
- fwrite
- fread
- fputs
- fclose
- fflush
- fseek

It is worth noting that, at the moment, none of the stdio API is thread
safe. I've also added wrappers for stdin, stdout and stderr in order to
be compliant with libc behavior.

Signed-off-by: Louis Solofrizzo <lsolofrizzo@online.net>