ukl: add printf and family
This patch introduces printf in the ukl library. The following function
have been added:
- puts
- vfprintf
- printf
- vfprintf
- vprintf

All snprintf functions and like are already present in the kernel.
There's also the beginning of stdlib, with basic support for wchar_t,
and limits.h.

It is worth noting that printf does _not_ handle float and double, since
there are no FPU in the kernel.

Signed-off-by: Louis Solofrizzo <lsolofrizzo@online.net>