ukl: implementation of the hybrid unikernel
This patches introduces a new boot param: ukl-hybrid. When this
parameter is passed to the kernel, the UKL does not hand indefinetely
after launch, and instead launch the init process in userspace like a
normal boot. The idea behind this is to keep network conf / monitoring /
debug / access / etc in userspace, in order not to rewrite / port the
entire GNU userspace in kernel. The 'ukl' option still launch the UKL
without starting the userspace at all.

Signed-off-by: Louis Solofrizzo <lsolofrizzo@online.net>